Additional work on uninstall in support of bugzilla 191018 - Initial support for uninstall. Includes fixes, though possibly not the final solutions, for 191610 - Can't uninstall org.eclipse.osgi due to EquinoxBundlesState.composeNewState() and 191612 - Once equinox.simpleconfigurator is uninstalled, bundles.txt is no longer updated.
I have tested install and complete uninstall of an SDK. The result is relatively clean, removing the config.ini file, the org.eclipse.equinox.simpleconfigurator directory and its contents, and the eclipse .ini file. No attempt is yet made to remove the launcher .exe file or and of the configuration subdirectories left by executing the installed SDK.

The partial uninstall scenario - for example, installing an SDK then extending it by, for example, installing CDT, then removing CDT does not currently work. The dependency computation in the directory is going wrong and the sdk bundles will be removed as well. I will look into this problem tomorrow.