Fix for bug 208003 - Many versions of "self" capability in repository.
Change InstallableUnit and InstallableUnitFragment so that do implicitly add the self and fragment capabilities. Changed the metadata generator and test to always explicitly add the appropriate self and fragment capabilities. Did some restructuring in the automated tests so that most creation of IUs use utilities that handle the capabilities appropriately.

Reorganized the xml write/parsers so that reading and writing of IUs can share code (in metadata repositories and the metadata cache).

Miscellaneous cleanup related to the xstream removal.