Initial commit of metadata generator support for the IUs and artifacts for eclipse features. Two IUs are now generated per feature: the group IU which has been renamed to id.featureGroup (e.g org.eclipse.platform.source.featureGroup) and the IU which actually holds the feature artifact, id.featureIU (e.g. org.eclipse.platform.source.featureIU). A filter of "eclipse.p2.install.features=true" is applied to the installation and configuration of the artifacts. This property is now set in the profile, but we need to better define the Update Compatibility use case to determine the right way to get the filter set. Currently a vm arg (system property) -Dorg.eclipse.p2.update.compatibility=true  is used to  feature installation by setting the value of the "eclipse.p2.install.features" property in a new profile to be the value of the system property.
The installation and configuration of features is done through IPlatformConfiguration, so the org.eclipse.platform.configurator plug-in must be installed.

The support for installing features is still work in progress. Currently when features are installed in a p2 provisioned SDK, the provisioned eclipse starts but gets system errors because the ConfigurationActivator in o.e.platform.configurator fails to start with a bundle exception. I have tried several versions of the resulting platform.xml and debugged through the start method of the activator, but have not discovered what is causing the failure. Eyes more familiar with the internals of update may be more successful.