Work on bug 217100:
The code in Profile was an early hack since we hadn't made decisions about the
level of update compatibility.

Talked with John about the right solution and we arrived at two conclusions:

1). We should use a separate profile property as the filter for feature install
since it is a general case of selection (like source or other random content
selections). The environment property should be reserved for os, ws, arch, nl.

2). The decision about selection filters (like feature or source install)
should be made by the creator of a profile and passed into profile construction
as properties.

John is going to make the changes for the installer. I am going to fix Profile
and the Director App.

I eliminated the hack in Profile.  The aguments to DirectorApp can now include -profileProperties  or -props, e.g.

-props  org.eclipse.p2.installFeatures=true,foo=bar