Fix for bug 234278: P2 Eclipse when installed into two sites does not come up after reconciliation.
Added eclipse.core.runtime and dependencies to bootstrap

Needed to have org.eclipse.core.runtime bundle in the bootstrap application and
started in config.ini.  Problem was masked by reconcile causing registry and
application service to be loaded when there were changes in platform.xml that
needed to be reconciled.

This is a non-code change.