334519 Added parameter for selecting JRE profile in product publisher application
Introducing a new parameter for the product publisher application to
read the java profile from an external file. The value of the specified
location can be either a single file or a JRE folder with a .profile
file. In the latter case, the whole JRE folder is zipped and uploaded as
binary artifact to p2 repository.
Also, fixed reading the versions of the packages, described in the java
profile.

Bug 334519 - [publisher] Erroneous behavior in product publisher when handling the java profile