Bug 349628 - No update proposed when there are repository with errors
When a repository fails to load, we continue the update check.  All the
repositories that fail to load are 'accumulated' and shown to the user,
but the update check proceeds with the available repositories.

This commit changes the visibility of a method in an API class from
private to protected. However, this class is marked as 'noextend' so
really shouldn't be considered new API.