Bug 384646: Improves the performance of the install software filter
This changeset includes two approaches to fix the performance problems
in the p2 Install Software Dialog:

1. The children of a category are now cached. This ensures that the tree
is only built once.

2. When computing the members of a category, we now do a 'piped query'.
Instead of searching all IUs in a repo to see if they are in a category,
we limit ourselves to only IUs which should be shown. This was available
from the policy as a "top level query".