Bug 423538 - Format IU list output of P2 director
By default the -list command only prints id and version of an IU,
which is not always sufficient.  For example, a UI might want to show
the name of IUs instead of their ID to the user.

This patch tries to accommodate this by introducing format strings
through a new -listFormat <formatString> command.  Using the ${property}
syntax one can access all properties of an IU in the format string.  ID
and version are supported through hard-wired ${id} and ${version}
expressions.

Existing behavior of -list (without -listFormat) is preserved.

Change-Id: I9c32f6040b0a75b450f230338ca612939b8765e3
Signed-off-by: Christian Georgi <christian.georgi@sap.com>