bug 421935: Extend simpleconfigurator to read .info files from many locations.
Locations should be passed to the simpleconfigurator via the
p2.fragments java property and separated using ",".

Each location may contain multiple fragments. A fragment consist of a
directory shaped as:
+ extension-parent-dir
  + extension.info
  + artifacts.xml
  + content.xml
  + plugins/*
  + features/*
Or a link file (f.e. extension.link) which contains a single java
property ("link") pointing to another location in the system.

The *.info file may contain relative or absolute paths to plugins.
Relative paths will be resolved against the extension root.

The simpleconfigurator, if configured with extensions, will attempt to
load them all at startup, duplicates will be removed.

It is the responsibility of the fragment provider to ensure that all
fragments will lead to a valid Eclipse installation, AS NO VALIDATION OR
INTEGRITY CHECK IS PERFORMED. 
Bundles that can't be loaded due to unsatisfied constraints will result
in errors handled by OSGi (written to the .log in the configuration
area).


Change-Id: Ib2689c072f3149b29ed77797d7df7b209d9abde7
Signed-off-by: Krzysztof Daniel <kdaniel@redhat.com>