Bug 422054: Create profile from existing bundles.info
I've hijacked the bug for the purpose of something that is remotely
connected to the problem.

Previous commit extends simpleconfigurator to load bundles from
different locations, but such an extension breaks p2, as the content of
OSGi application no longer corresponds to profile.

In order to fix that, following things are done:
0. A fragment must contain valid P2 metadata next to the .info file.
1. User profile is dropped when there is master configuration change.
Also, when fragments are changed.
2. When a new surrogate profile is created, P2 will add to the master
profile all the units that are found in the fragments repos. Therefore
new user profile will contain all the units that were installed by the
simpleconfigurator. Fragment repos must be runnable.
3. The director application must be configured with extensions if it is
supposed to work.

Change-Id: I6e857ea51dd32ae7fab39d9c39bec8a91eb203b7
Signed-off-by: Krzysztof Daniel <kdaniel@redhat.com>