Bug 427148 - Unable to update from one version to the next,
On Windows only - can not remove eclipse.exe

When Eclipse is installed on a different volume than the users home
directory, an update of the eclipse.exe fails. This is because the exe
is locked as a running process, and cannot be 'moved' to the backup
store because it exists on a different volume. To fix this we rename the
file in-place. This moves the exe 'out-of-the-way' so it can be updated.

This only works for files named eclipse.exe or Eclipse.exe

For more information, see:
http://bugs.eclipse.org/427148
http://eclipsesource.com/blogs/2014/05/16/read-the-javadoc-then-test-your-assumptions/

Change-Id: Ic3b6a633aa0a25d84e0219934418c5aa245c666d