Bug 466249 - Propagate more exceptions to ease problem analysis
We often get problem reports with stack traces that "end" somewhere in 
p2 even though it's obvious from looking at these places that there was 
a causing exception. These causing exceptions are often not propagated 
with the new exception that p2 creates and throws. I'm not sure if 
that's on purpose or just an oversight. I've prepared a proposal to 
enhance these places and make it easier to analyze the problems that 
users report.


Change-Id: Iaa53448c53c18301113b42dbe80558eccec49e8e
Signed-off-by: Eike Stepper <stepper@esc-net.de>