Bug 493193 - MirrorApplication should explicitly use MD5 artifact comparator
The only place that documents the usage of MD5 comparator in
MirrorApplication is Equinox p2 Repository Mirroring wiki page [1]:

  Adding the argument -comparator <comparator ID> specifies the
  mirroring application should use an Artifact Comparator with an ID of
  "comparator ID" to compare artifact descriptors. The mirroring
  application uses the "MD5 Comparator" to compare the MD5 hash property
  of the artifact descriptors if no comparator is defined.

If no comparator is defined, ArtifactComparatorFactory's
getArtifactComparator(String) will be finally called with null as a
comparator ID. It's an ArtifactComparatorFactory's implementation detail
that it will return the first configured one. MD5 comparator is the only
one available now, but this may change and break MirrorApplication's
expectation.

To avoid such breakage, MirrorApplication declares an explicit
dependency on MD5 comparator. So, when '-compare' CLI parameter is true
but '-comparator' CLI parameter is missing, it will request MD5 artifact
comparator to be used.

Also improve testBaselineCompareUsingMD5Comparator() and
testCompareUsingMD5Comparator(). These two tests should:
- not only set 'compare' to true but also set a desired comparator's id
  explicitly
- run twice, for an empty comparator id and MD5.

To not copy-n-paste them, use @Theory [2]. To make theories work, run
with Theories test runner.

[1] https://wiki.eclipse.org/Equinox_p2_Repository_Mirroring
[2] https://github.com/junit-team/junit4/wiki/Theories

Change-Id: I5adda382c2824b6488af3716fc645cb6eff14b94
Signed-off-by: Mykola Nikishov <mn@mn.com.ua>