Bug 495726 - API: InstallWizardPage ui elements do not respect policy
P2 UI Wizard pages do not respect provisioning ui which is passed to
them when
it comes to querying elements, because IUElementListRoot which is the
root of all elements does not 
extend QueriedElement so getProvisioningUI() on the elments defaults 
to ProvisioningUI.getDefaultUI()
Fix is to add an API to accept ProvisioningUI as an argument. All places
that were creating IUElementListRoot now pass appropriate ui to the
constructor. For backward compatibility if ui is not set or set to null
it will return default UI as before.

Change-Id: I1d8d87339812c874eeee7127e762731b600cb4bf
Signed-off-by: Alena Laskavaia <elaskavaia.cdt@gmail.com>