Bug 486279 - Changed missing entry status to warning
Converted the errors caused by missing entries in bundles.info to 
warnings. This keeps the actions backward compatible. For some of the 
actions is makes sense as well. E.g. uninstall should be idempotent

Refactored UninstallBundleAction to relay on Util.findBundleInfo() to
find the entry in bundles.info, rather than to try to re-construct the
BundleInfo from other data and use Object.equals()

Change-Id: I902a5b209b1ff6a764e6c8f0569926c97ed8b3c5
Signed-off-by: Todor Boev <rinsvind@gmail.com>