Bug 490896: Director app handles all result status codes.
When the provisioning operation completes with WARNING or INFO the 
director command line application reports failure (equinox returns 
non-zero execution code). At the same time P2 has in fact performed 
changes to the installation and these were not reverted as would happen 
on an ERROR IStatus. The director should not report failure when 
provisioning completes with INFO or WARNING IStatus.

Change-Id: I9a683174ffe9e49aa4456b4aa5134f9067ecc7a7
Signed-off-by: Todor Boev <rinsvind@gmail.com>