Bug 313553 - Unentangle ProvidedCapability members and attributes
The members are the generic way for p2 queries to describe getters
on objects. The capability attributes then participate in queries
as the "attributes" member of a ProvidedCapability object.

Fixed the IProvidedCapability to follow the OSGi convention that the
attributes contain the name of the capability under a key equal to the
capability namespace. This is cleaner than to introduce a new "name"
attribute for the p2 name that can conflict with an incoming OSGi "name"
attribute.

Fixed the version of IProvidedCapability to follow the OSGi convention
that capabilities with a version have a "version" attribute with a value
of type an OSGi version object.

Change-Id: Ic9b77c2d103216141035dc3fb5861ca99a3ddccd
Signed-off-by: Todor Boev <rinsvind@gmail.com>