Bug 531606 - RequiredCapability introduced binary-incompatible changes
p2 tries to not change internals (due to bad downstreams using
internals widely) but these are internals after all and considering
the churn with capabilities work and current stable state, we are not
going to do any work here.

Clients would have to adapt to the changes when moving to Photon p2.

To communicate this, explicitly mark RequiredCapability as not
intended to be referenced by clients.

Change-Id: I4bddb36f13178ec07752ee92b6bacaca26370bf2
Signed-off-by: Mykola Nikishov <mn@mn.com.ua>