Bug 532118 - Reduce number of calls to java.security.MessageDigest.update(byte)
MD5Verifier and ChecksumVerifier are responsible for calculating
artifact's checksum and must extend ProcessingStep to fulfill API
requirements. ProcessingStep is effectively an unbuffered OutputStream
with a method to process a single byte of artifact's data.

In respect to MessageDigest, this means updating digest with every new
single byte of data. So, for 16 Kb of data, both classes will call
MessageDigest's update(byte) 16K times while only 1 call is enough.

Introduce MessageDigestProcessingStep that accumulates incoming bytes
in the internal buffer and immediately forwards them to the
destination stream. As soon as buffer is full, it will update
MessageDigest.

AbstractBufferingStep could be used as a base class instead but it is
too risky to adapt it. java.nio.ByteBuffer fits the bill almost
ideally. One caveat is it communicates 'buffer is full' by throwing
BufferOverflowException.

There are no specific reasons for buffer's size of 16K except an
assumption that it should not cause any problems: installing 1000
artifacts with 2 checksums per artifact would increase memory
footprint for about 3200K.

Similar approach has been used already [1], [2], claiming 5x speedup.

This should make [3] less of a problem.

[1] 5b2f061cd8ecf96a37783657e7ee7a0110c9d26d
[2] Bug 405716
[3] Bug 532036

Change-Id: I1ba0d36267b3f07103c22b93f92c41765e33c8cc
Signed-off-by: Mykola Nikishov <mn@mn.com.ua>