Bug 532118 - Eliminate exception when checking if buffer is full in MessageDigestProcessingStep
There is common wisdom behind the claim that throwing exception may
hurt performance but I have no real numbers to say it will have real
impact in this case.

However, in this case it is easy to avoid exception by explicitly
checking if the buffer is full before putting new data into it.

Change-Id: Ifbd178aff4f943c31772aa357ee8369efcee558f
Signed-off-by: Mykola Nikishov <mn@mn.com.ua>