Bug 533117 - setRepositoryProperty: fixed regression when value is null
A null value should remove the entry from the repositoryProperties. But
when repositoryProperties is not initialized a null value was actually
added for the key.
Changed conditions to first prove wether value is null and only remove
the property, when field repositoryProperties is initialized. Set field
to null when the last entry was removed.

Change-Id: I766094f09d24bcf3c0da093202724ab5c2f0e417
Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>