Bug 512407 - Let's dump 'ie.wombat.jbdiff' and related from Git repository
Never have been published in our repo so safe to remove. Removed parts
that used it but these bundles are not in the repo also so safe again.
It is sure that if/when jbdiff is really used it should be added to
Orbit and used proper instead of keeping it's source in our repo.

Change-Id: Iac3cccd08d3019d70d1d18632a761d20bad444b8
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>