Bug 541444 - source.info contains old/duplicate bundles after platform upgrade
This change takes advantage of the fact that simpleconfigurator
has already done the checking to see if the shared installation
has been updated and set a system property accordingly. We then
use the presence of the system property to determine whether to
ignore the user configuration or not.

Ignoring the user configuration means the source.info will be re
-generated from scratch using the bundles in the updated shared
configuration, then the migration wizard will re-insert any third
-party bundles as appropriate.

This avoids the possibility that there will be out-of-date
entries in the source.info in the user configuration area.

Change-Id: Ia59ce22133c928fe23b72c65a6034c5267f38f1b
Signed-off-by: Mat Booth <mat.booth@redhat.com>