Use jdk 5 for-each loop
Replace simple uses of Iterator with a corresponding for-loop. Also add
missing braces on loops as necessary.

toString() on array

toString() implementation for arrays does not convert array contents to
String, rather it will print array's type and hash code (defined as
identity hash code). To get string representation of contents of the
array, the array could be for example wrapped into Arrays.asList(), as
Collections produce content representation in their toString().

Change-Id: I5db7c1cb516f25cb8410278d1d687862de471c6a
Signed-off-by: Carsten Hammer <carsten.hammer@t-online.de>