Bug 545045 - Cancellation.testCancellation fails on Windows
This test only runs on Windows but I think it should be able to run on
any platform so I've enabled that to see if this is the case. Hopefully
this will run in the Gerrit verification job, but then it might fail
unless it's really using the verification application's changes built in
that Gerrit job. We'll see.

The failure I'm pretty sure the failure is caused by the fact that the
saving the profile preference is done by a job with a delayed schedule
and that the verification application terminates before the save
actually happens, so I've added join.  I'm pretty sure of this because
if I remote debug the launched verification application set breakpoints
such that the preferences are really saved, the test runs without
failure.

But it's hard to test this without committing because it appears the
verification application comes from the zip file being tested, not from
the workspace version that I've changed.

Change-Id: Ic30c06cb4c44ac75842c863d6f7a0360d2137555
Signed-off-by: Ed Merks <ed.merks@gmail.com>