Bug 563485 - Use TextMatcher instead of StringMatcher
TextMatcher is a newer and leaner re-implementation of StringMatcher
providing exactly the same functionality. TextMatcher can be used as
a drop-in replacement; StringMatcher will be removed once this
replacement is done.

Note that both StringMatcher and TextMatcher are *internal* classes
in org.eclipse.ui.workbench and are available here via the re-export
in org.eclipse.ui.

Change-Id: I4d2f88ce8ad7eebd6a69e6921e72d2c457686667
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>