Bug 565066 - Force reresolve of all bundles on install/uninstall
In order to maximize the ability to select a good resolution after
incrementally installing or uninstalling bundles all existing bundles
are forced to re-resolve with the new set of available bundles.  This
forces the framework to forget its previous decisions for resolution
wiring so that it may make better choices in order to resolve the newly
installed set of bundles.

Change-Id: Ic28569f7dc35407261e7a39087817e572d7e8c8d
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>