Bug: Possible null pointer dereference of profileDirectories in org.eclipse.equinox.internal.p2.engine.SimpleProfileRegistry.restore()
There is a branch of statement that, if executed, guarantees that a null
value will be dereferenced, which would generate a NullPointerException
when the code is executed. Of course, the problem might be that the
branch or statement is infeasible and that the null pointer exception
can't ever be executed; deciding that is beyond the ability of SpotBugs.

Rank: Scary (8), confidence: Normal
Pattern: NP_NULL_ON_SOME_PATH
Type: NP, Category: CORRECTNESS (Correctness)

see https://bugs.eclipse.org/bugs/show_bug.cgi?id=336435

Change-Id: Ib73e1a65f6428beea9b3f6aa6b88b4903f86a9ec
Signed-off-by: Carsten Hammer <carsten.hammer@t-online.de>