[dogfooding][cleanup] Use new StringBuffer -> StringBuilder cleanup for p2
To ensure that the new JDT cleanup works correctly for clients we should
run it also on our own code base. This help to ensure that the cleanup
works fine and helps us to cleanup our own code base.

From the commit message of the cleanup:

- create a cleanup to replace usage of StringBuffer with
  StringBuilder.  The two are equivalent, but StringBuffer is
  thread-safe and synchronized which makes it slower than
  StringBuilder which was introduced in Java 1.5.


To avoid API changes, we should only use the local variable change.

Change-Id: Iab5520fa3bf21f5374cc70de2a8fd4b63790027a
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>
Reviewed-on: https://git.eclipse.org/r/c/equinox/rt.equinox.p2/+/182677
Tested-by: Equinox Bot <equinox-bot@eclipse.org>
Reviewed-by: Alexander Kurtakov <akurtako@redhat.com>