Bug 576705 - Declare some PGP keys as trusted in p2 metadata
This allows to build a product/profile or install metadata that define a
pgp.trustedPublicKeys property containing public keys to trust.

After verifying that signatures are correct when installing an artifact
(was bug 570907), deciding of whether to consider an artifact as trusted
or not will check whether one of the signing keys is already installed
trusted.
If no signing key is trusted, user is prompted for approval in order to
continue installation, as already implemented with bug 572816

Change-Id: Ie17ecca23de8186845ec943e63abb4dd0d141836
Reviewed-on: https://git.eclipse.org/r/c/equinox/rt.equinox.p2/+/186727
Tested-by: Equinox Bot <equinox-bot@eclipse.org>
Reviewed-by: Mickael Istria <mistria@redhat.com>