Bug 578091 - The TrustCertificateDialog's Details... button is never enabled
In the chain viewer,  set the selection when the input changes so that
the details button enables according to that selection and properly
update the Details... button when the selection is a tree node with a
certificate.

Change the menu item to copy the fingerprint from "Copy fingerprint" to
"Copy Fingerprint" (because it should be title case).

Use SWT.FULL_SELECTION for the checkbox table viewer so that selecting
in any column selects the row.

Properly enable/disable the Trust button depending on whether at least
one item is check marked (rather than based on whether the selected item
is check marked).

Change-Id: I932fbba94ea93691317845efd7ab454d509bd87d
Signed-off-by: Ed Merks <ed.merks@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/equinox/rt.equinox.p2/+/189410
Tested-by: Equinox Bot <equinox-bot@eclipse.org>