Bug 578161 - The CertificateChecker uses the wrong profile
Provide support to be able to set the profile used by the
CertificateChecker, much like it supports adding the artifacts
descriptor to add.

Use this in the CheckTrust phase so that the CertificateChecker operates
on the profile being provisioned.

Modify downstream uses of the CertificateChecker to set the appropriate
profile for that usage context.

Change-Id: Ie2da8e1b137f8540ce40f8dd737ffc938a47194d
Reviewed-on: https://git.eclipse.org/r/c/equinox/rt.equinox.p2/+/189475
Tested-by: Equinox Bot <equinox-bot@eclipse.org>
Tested-by: Ed Merks <ed.merks@gmail.com>
Reviewed-by: Ed Merks <ed.merks@gmail.com>
Reviewed-by: Mickael Istria <mistria@redhat.com>