Bug 575541 - [pgp] P2 should not fail completely if a public key is not found
Ignore signatures for which no corresponding key can be found.

If all signatures are ignored, the artifact will be effectively treated
the same as an unsigned artifact.

In the unsigned artifact details presented to the user, include the key
ID(s) of the ignored signature(s).

Update the test to expect the artifact request status to be OK.


Change-Id: If433a81144d2eafed51e12d38396e9ffa5b09787
Signed-off-by: Ed Merks <ed.merks@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/equinox/rt.equinox.p2/+/189588
Tested-by: Equinox Bot <equinox-bot@eclipse.org>