Bug 578232 - Improve the trust preference page
Differentiate between the keys trusted by the user and stored in profile
preferences versus the ones contributed by the extension point.  The
latter are not removable and should provide information to the user
about which bundle contributed those keys.

The preference page handles multiple selection better, enables the
Remove button only when at least one of the selected keys is contributed
from the preferences, and enables the Export button if one or more keys
are selected.

It uses  a TableColumnLayout so that the user can see all the
information more easily.

When keys are added, the newly added keys are selected so that you know
what you added.

The Restore Default button will reset the state to what it was when the
page was first opened.

The file dialogs that open remember their location in the dialog
settings so that it's easier to add/export multiple keys.  They default
to user.home folder.

The description message is made more concise.

The items are sorted for a more consistent presentation; items from the
preference store come first.

Change-Id: I3c520b40fb8940a2ed02dde0376f6d5535f50c45
Signed-off-by: Ed Merks <ed.merks@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/equinox/rt.equinox.p2/+/190452