Bug 578024 - Improve the TrustCertificateDialog
Use a SashForm so that relative sizes of the top table and the bottom
tree can be controlled by the user.

Use a wrapping label for the description text such that a longer
description doesn't make the dialog too wide.

Move the warning text about the name not being reliable from the column
label to the description.

Use a TableColumnLayout so that all the columns are visible without
scrolling.

Fix the problem with the false indication of expiration, i.e., when the
valid seconds is 0, that means there is no expiration.

Change-Id: Ib3f41145174a066d0c48c4e1cc91ca580f89b510
Signed-off-by: Ed Merks <ed.merks@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/equinox/rt.equinox.p2/+/190485