Bug 578024 - Improve the TrustCertificateDialog
Introduce
org.eclipse.equinox.p2.repository.artifact.spi.IArtifactUIServices
optionally implemented by UIServices implementations for enriched
information that includes information about the artifacts associated
with certificates and keys. Ensure that the new interface can delegate
to older UIServices implementations.

Upgrade CertificateChecker to compute the more detailed information and
that it uses IArtifactUIServices when available, but also works properly
when it's not available.

Update ValidationDialogServiceUI so that it implements
IArtifactUIServices and uses the enhanced information when available.

Completely overhaul the TrustCertificateDialog implementation.

Change-Id: I268bc7176af3eeb98b1d38a47b92d525f9b9f285
Signed-off-by: Ed Merks <ed.merks@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/equinox/rt.equinox.p2/+/190518