Bug 578024 - Improve the TrustCertificateDialog
Enhance UIServices.TrustInfo to include a boolean trustAlways() and
enhance TrustCertificateDialog to provide isTrustAlways() as well as
isRememberSelectedSigners().  The UIServices.TrustInfo's persistTrust()
was never properly/explicitly populated to reflect whether the user
actually wanted to remember the selected signers or simply wanted to
trust the signers for this one operation and be prompted again in the
future.  TrustCertificateDialog includes checkbox for populating
isRememberSelectedSigners() and isTrustAlways().  The former is
true/checked by default and of course the latter is false by default.
Also include a warning/confirmation about the implications of trusting
all contents always.

The result of TrustInfo.isTrustAlways is stored as a profile preference
via CertificateChecker.setTrustAlways(boolean) and that preference, via
CertificateChecker.isTrustAlways(), is respected by
CertificateChecker.checkCertificates.

The preference is available via TrustPreferencePage which also includes
a warning/confirmation when the user tries to enable it.

Change-Id: I02f34456f4099870136b54022990d5aa5a4cafb9
Signed-off-by: Ed Merks <ed.merks@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/equinox/rt.equinox.p2/+/190598