Fix p2.tests.reconciler product
With the latest changes in Equinox [1] o.e.osgi.util was replaced by the
corresponding 'original' OSGi-reference bundles that are
required+re-exported by o.e.osgi.util. Unfortunately the re-export does
not help in plug-in based products, but 'o.e.osgi.util' seems not to be
necessary so it is just removed.

[1] - https://github.com/eclipse-equinox/equinox.framework/issues/40