Issue #90 - Avoid refreshing simple configurator fragments
When a fragment is refreshed it causes the host to refresh.  This causes
a (temporary) deadlock because simple configurator refreshes bundles
while it is starting.  The deadlock breaks after a timeout but that
results in an incomplete refresh package operation.  This fix avoids
causing refresh of simple configurator fragments similar to the system
bundle fragments.